+++ src/common/enums/socket-service-name.enum.ts
@@ -1,0 +1,7 @@
+export enum SocketServiceName {
+  Mqtt = 'mqtt',
+  Websocket = 'websocket',
+  SocketIo = 'socketIo',
+  RpcStreaming = 'rpc_streaming',
+  KafkaStreaming = 'kafka_streaming',
+}