+++ src/common/interfaces/notifications/notifier.interface.ts
@@ -6,6 +6,17 @@ export interface SmsResponse {
   error?: string | undefined;
 }
 
+export interface SocketReport {
+  sent: boolean;
+  seen: boolean;
+}
+
+export interface SocketDelivery {
+  sent_message: string;
+  priority_level: string;
+  links?: string[];
+  reports?: Record<string, SocketReport>;
+}
 export interface SmsDelivery {
   reports: Record<string, SmsResponse & { user_id?: string }>;
   sent_message?: string;
@@ -13,15 +24,13 @@ export interface SmsDelivery {
 
 export interface NotifierDelivery {
   sms?: SmsDelivery;
+  socket?: SocketDelivery;
+
   mail?: {
     status: Record<string, unknown>[];
     sent_message?: string;
   };
   fcm?: {
-    status: Record<string, unknown>[];
-    sent_message?: string;
-  };
-  socket?: {
     status: Record<string, unknown>[];
     sent_message?: string;
   };
@@ -32,6 +41,20 @@ export interface NotifyByReason extends Base {
   value_args?: string[];
   metadata?: Record<string, unknown>;
   delivery?: NotifierDelivery;
+}
+export interface NotifyByCustom extends Base {
+  reason: string;
+  recipients: CustomSmsRecipients;
+  value_args?: string[];
+  metadata?: Record<string, unknown>;
+  delivery?: NotifierDelivery;
+}
+
+export interface CustomSmsRecipients {
+  alias: string;
+  phone_numbers?: string[];
+  subjects?: string[];
+  user_ids?: string[];
 }
 
 export type NotificationNotifierInterface = NotifyByReason;