+++ src/common/interfaces/workflow.interface.ts
@@ -1,9 +1,9 @@
 import { Base } from './base';
 import { WorkflowToken } from './workflows';
 
-export interface Workflow extends Base {
+export interface Workflow<T = unknown> extends Base {
   type: string;
-  data?: unknown;
+  data?: T;
   status: string;
   tokens: WorkflowToken[];
 }