+++ src/common/models/notifications/subs/options/sms-options.model.ts
@@ -1,8 +1,8 @@
-import { SmsFunctionality } from '../../../../enums';
+import { SmsProviderFunctionalityEnum } from '../../../../enums';
 import { Options } from './base-options.model';
 
 export class SmsOptions extends Options {
-  functionality!: SmsFunctionality;
+  functionality?: SmsProviderFunctionalityEnum;
   provider_lookup_template_name?: string;
   retry?: boolean;
 }