+++ src/common/models/notifications/template.model.ts
@@ -3,6 +3,7 @@ import { FcmOptions, MailOptions, SmsOptions, SocketOptions } from './subs';
 
 export class TemplateModel extends BaseModel<TemplateModel> {
   name!: string;
+  active?: boolean;
   description?: string;
   reason!: string;
   sms?: SmsOptions;