+++ CHANGELOG.md
@@ -7,6 +7,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 2.7.0 - 2023-09-05
+### Added
+- Notify by Custom Recipients
+
+### Changed
+- Notification (all modules) type (models, interfaces)
+
+### Fixed
+- Zones added to BaseModel
+- Generic Workflow interface to specify Data type
+
+## 2.6.1 - 2023-08-26
+### Fixed
+- Addressing Common and Infrastructure folders
+
 ## 2.6.0 - 2023-08-15
 
 ## 2.5.4 - 2023-08-12